third_party/WebKit/LayoutTests/fast/html/imports/import-dynamic-tree.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
description("This test ensures that script-added import lets its child blocks following scripts.");

window.jsTestIsAsync = true;
window.trace = [];

var linkEl = document.createElement("link");
linkEl.setAttribute("rel", "import");
linkEl.setAttribute("href", "resources/async-root.html");
document.head.appendChild(linkEl);
linkEl.onload = function(evt) {
    shouldBe("window.trace", "['START root', 'child1', 'child2', 'END root']");
    finishJSTest();
};

</script>
</body>
</html>
