third_party/WebKit/LayoutTests/fast/html/imports/import-dynamic-tree-expected.txt

This test ensures that script-added import lets its child blocks following scripts.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.trace is ['START root', 'child1', 'child2', 'END root']
PASS successfullyParsed is true

TEST COMPLETE
